+++ example/lib/main.dart
@@ -10,31 +10,34 @@ import 'page/android_system_setting.dart';
 
 Future<void> main() async {
   isBeta = true;
-  await Global().setConfig(GlobalConfig(
-      mainColor: Colors.purple.shade900,
-      textColor: TextColor(
-          largeColor: const Color(0xFF292929),
-          veryLargeColor: const Color(0xFF292929),
-          defaultColor: const Color(0xFF292929),
-          styleColor: const Color(0xFF292929),
-          smallColor: const Color(0x804D4D4D)),
-      loadingBuilder: (BaseLoading loading) => Container(
-          width: loading.size * 2,
-          height: loading.size * 2,
-          decoration: BoxDecoration(
-              color: UCS.black, borderRadius: BorderRadius.circular(10)),
-          child: const BaseLoading(
-              color: Colors.white, style: SpinKitStyle.fadingCircle)),
-      betaApi: '这是设置测试Api',
-      releaseApi: '这里设置发布版Api',
-      toastOptions: const ToastOptions(ignoring: false)));
-  BasePackageInfo().initialize();
+  await Global().setConfig(
+      GlobalConfig(
+          mainColor: Colors.purple.shade900,
+          loadingBuilder: (BaseLoading loading) => Container(
+              width: loading.size * 2,
+              height: loading.size * 2,
+              decoration: BoxDecoration(
+                  color: UCS.black, borderRadius: BorderRadius.circular(10)),
+              child: const BaseLoading(
+                  color: Colors.white, style: SpinKitStyle.fadingCircle)),
+          betaApi: '这是设置测试Api',
+          releaseApi: '这里设置发布版Api',
+          toastOptions: const ToastOptions(ignoring: false)),
+      windowOptions: WindowOptions(
+          size: WindowsSize.iPhone5P8.value,
+          minimumSize: WindowsSize.iPhone4P7.value,
+          maximumSize: WindowsSize.iPhone6P1.value,
+          center: true,
+          backgroundColor: Colors.transparent,
+          skipTaskbar: false,
+          title: 'Universally'));
+  PackageInfoPlus().initialize();
   runApp(BaseApp(
       title: 'Universally',
       providers: [ChangeNotifierProvider(create: (_) => AppState())],
       home: const HomePage(),
       initState: (context) async {
-        BaseConnectivity().addListener((status, result) async {
+        ConnectivityPlus().addListener((status, result) async {
           switch (result) {
             case ConnectivityResult.wifi:
               showToast('use wifi');
@@ -60,7 +63,7 @@ Future<void> main() async {
           }
           return true;
         });
-        BaseConnectivity().subscription(
+        ConnectivityPlus().subscription(
             alertUnavailableNetwork: (status, result) =>
                 alertOnlyMessage('Network Unavailable'));
       }));
@@ -78,11 +81,6 @@ class HomePage extends StatelessWidget {
         enableDoubleClickExit: true,
         child: Wrap(alignment: WrapAlignment.center, children: [
           Button(onTap: () => push(const ComponentPage()), text: 'Component'),
-          Button(
-              onTap: () {
-                UrlLauncher().openUrl('tel:10086');
-              },
-              text: 'Call Phone'),
           Button(onTap: () => push(const GifPage()), text: 'Gif'),
           Button(child: const SwitchApiButton(color: Colors.white)),
           Button(onTap: () => push(const TextFieldPage()), text: 'TextField'),
@@ -98,6 +96,21 @@ class HomePage extends StatelessWidget {
               text: 'showDoubleChooseAlert'),
           Button(onTap: showLoading, text: 'showLoading'),
           Button(
+              onTap: () {
+                push(const SpinKitPage());
+              },
+              text: 'SpinKit'),
+        ]));
+  }
+
+  List<Widget> get buildMobile => !isWeb && isMobile
+      ? [
+          Button(
+              onTap: () {
+                UrlLauncher().openUrl('tel:10086');
+              },
+              text: 'Call Phone'),
+          Button(
               onTap: () async {
                 final res = await getPermission(
                     Permission.requestInstallPackages,
@@ -105,13 +118,6 @@ class HomePage extends StatelessWidget {
                 showToast(res.toString());
               },
               text: 'requestInstallPackages'),
-          Button(
-              onTap: () async {
-                final res = await getPermission(Permission.camera,
-                    alert: '本服务需要访问您的“相机”，以修改头像或上传图片');
-                showToast(res.toString());
-              },
-              text: 'getPermission'),
           Button(
               onTap: () async {
                 final res = await getPermissions([
@@ -123,12 +129,24 @@ class HomePage extends StatelessWidget {
               },
               text: 'getPermissions'),
           Button(
+              onTap: () async {
+                final res = await getPermission(Permission.camera,
+                    alert: '本服务需要访问您的“相机”，以修改头像或上传图片');
+                showToast(res.toString());
+              },
+              text: 'getPermission'),
+          Button(
               onTap: () {
                 UrlLauncher().openAppStore(
                     packageName: 'com.tencent.mobileqq',
                     appId: isIOS ? '444934666' : '451108668');
               },
               text: 'openAppStore'),
+        ]
+      : [];
+
+  List<Widget> get buildAndroid => !isWeb && isAndroid
+      ? [
           Button(
               onTap: () async {
                 final result = await UrlLauncher().isInstalledApp(
@@ -142,18 +160,20 @@ class HomePage extends StatelessWidget {
                 push(const AndroidSystemSettingPage());
               },
               text: 'AndroidSystemSetting'),
+        ]
+      : [];
+
+  List<Widget> get buildIOS => !isWeb && isIOS
+      ? [
           Button(
               onTap: () {
                 UrlLauncher().openUrl(IOSSettingUrl.notifications.value);
               },
               text: 'IOSSystemSetting'),
-          Button(
-              onTap: () {
-                push(const SpinKitPage());
-              },
-              text: 'SpinKit'),
-        ]));
-  }
+        ]
+      : [];
+
+  List<Widget> get buildDesktop => !isWeb && isDesktop ? [] : [];
 }
 
 class Button extends Universal {