+++ lib/component/material_app.dart
@@ -178,14 +178,7 @@ class _BaseAppState extends State<BaseApp> with WidgetsBindingObserver {
     }
     super.initState();
     addObserver(this);
-    addPostFrameCallback((_) async {
-      widget.initState?.call(context);
-      if (isDebug && isDesktop) {
-        await Curiosity().desktop.focus();
-        final state = await Curiosity().desktop.setSizeTo6P1();
-        '桌面端限制宽高 $state'.log();
-      }
-    });
+    widget.initState?.call(context);
   }
 
   @override
@@ -264,7 +257,7 @@ class _BaseAppState extends State<BaseApp> with WidgetsBindingObserver {
 
   @override
   void dispose() {
-    BaseConnectivity().dispose();
+    ConnectivityPlus().dispose();
     removeObserver(this);
     super.dispose();
     widget.dispose?.call();