+++ scripts/windows.sh
@@ -1,18 +1,16 @@
 #!/usr/bin/env bash
 
 echo ""
-echo "========== 💪 开始打包iOS 💪 =========="
+echo "========== 💪 开始打包windows 💪 =========="
 
 # -n appName app名字  [appName]
 # -c channel 渠道名  [channel]
-# -m exportMethodName  [ad-hoc,app-store(default),development,enterprise]
 # -b buildType buildType [release,profile,debug]
 # -p mainPath
 
 outputName=""
 mainPath="lib/main.dart"
-channel="ios"
-exportMethodName="app-store"
+channel="windows"
 buildType="release"
 
 while getopts ":n:c:m:b:p:" optname; do
@@ -22,9 +20,6 @@ while getopts ":n:c:m:b:p:" optname; do
     ;;
   "c")
     channel="$OPTARG"
-    ;;
-  "m")
-    exportMethodName="$OPTARG"
     ;;
   "b")
     buildType="$OPTARG"
@@ -46,31 +41,30 @@ dartDefine=" --dart-define channel=$channel"
 if [ "" == "$channel" ]; then
   dartDefine=""
 fi
-exportMethod=" --export-method $exportMethodName"
 
 echo "
  -------------------------------------------------------------
- |    ios: $version,
+ |    windows: $version,
  |    outputName: $outputName,
  |    channel: $channel,
- |    exportMethod: $exportMethod,
  |    buildType: $buildType,
  |    mainPath: $mainPath
  --------------------------------------------------------------"
 
-echo "flutter build ipa --$buildType --analyze-size$exportMethod$dartDefine -t $mainPath"
+echo "flutter build windows --$buildType --analyze-size$dartDefine -t $mainPath"
 # shellcheck disable=SC2086
-flutter build ipa --$buildType --analyze-size$exportMethod$dartDefine -t $mainPath
+flutter build windows --$buildType --analyze-size$dartDefine -t $mainPath
 
-outputDir="app/ios/$buildType/"
+outputDir="app/windows/$buildType/"
 
 mkdir -p "$outputDir"
 
 echo ""
-echo "😄 打包 $channel 的 $exportMethodName $buildType 已完成 😄"
+echo "😄 打包 $channel 的 $buildType 已完成 😄"
 
-mv "./build/ios/ipa/$outputName.ipa" "./$outputDir$outputName-$channel-$exportMethodName-v$version-$(date "+%Y%m%d%H%M").ipa"
+# shellcheck disable=SC2140
+mv "./build/windows/x64/runner/$buildType" "./$outputDir$outputName-$channel-v$version-$(date "+%Y%m%d%H%M")/"
 
 echo ""
-echo "========== 💪 iOS打包完成 💪 =========="
+echo "========== 💪 windows 打包完成 💪 =========="
 echo ""