+++ CHANGELOG.md
@@ -1,4 +1,26 @@
-## 6.2.1
+## 6.4.3
+
+* Modify the `maxLines` and `minLines` of the `BaseTextField` and add the `TextEditingController` extension method
+
+## 6.4.1
+
+* Update dependency
+* example Supports windows
+
+## 6.4.0
+
+* Removed `DesktopWindowsSize`, and Added `WindowsSize`.
+* Removed `Curiosity().desktop`, and Added `window_manager`.
+* `Global().setConfig()` adds `windowOptions`
+* The scripts directory adds macos and windows support
+
+## 6.3.1
+
+* Change `BasePackage` to `PackageInfoPlus`
+* Change `BaseConnectivity` to `ConnectivityPlus`
+* Change `BaseDeviceInfo` to `DeviceInfoPlus`
+
+## 6.2.2
 
 * Migrate to 3.16.0
 
@@ -57,8 +79,10 @@
 
 ## 3.5.2
 
-* In `BasicAppBar`, change the `text` to `titleText` and change the `title` type to `Widget`, Change hasLeading to enableLeading
-* In `BasicScaffold`, add `appBarTitleText` and change the `appBarTitle` type to `Widget`, Change hasLeading to enableLeading
+* In `BasicAppBar`, change the `text` to `titleText` and change the `title` type to `Widget`, Change
+  hasLeading to enableLeading
+* In `BasicScaffold`, add `appBarTitleText` and change the `appBarTitle` type to `Widget`, Change
+  hasLeading to enableLeading
 * `UserPrivacyCheckbox` adds `fontSize`
 
 ## 3.5.1
@@ -121,7 +145,8 @@
 
 ## 1.7.0
 
-* Remove `showNetworkToast`, `alertNotNetwork`, `onConnectivityChanged` from `BasicApp()`,Please use `BasicConnectivity()`
+* Remove `showNetworkToast`, `alertNotNetwork`, `onConnectivityChanged` from `BasicApp()`,Please
+  use `BasicConnectivity()`
 
 ## 1.6.2+2
 
@@ -138,7 +163,8 @@
 ## 1.5.0+1
 
 * Remove appPath from `ProjectConfig()`,
-* Add cachePath to `ProjectConfig()`,If you do not set cachePath, the default cache path is automatically used
+* Add cachePath to `ProjectConfig()`,If you do not set cachePath, the default cache path is
+  automatically used
 
 ## 1.3.3